requirements.txt
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg==3.3.3
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
#g